Moved statistics things out of astats, docs to more appropriate place
